backend/AnalysisServer/requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==3.2.12
django-cors-headers==3.11.0
djangorestframework==3.13.1
dnspython==2.2.1
ifaddr==0.1.7
joblib==1.1.0
numpy==1.21.5
pandas==1.3.5
py-eureka-client==0.10.5
pycodestyle==2.8.0
pymongo==4.0.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
sqlparse==0.4.2
threadpoolctl==3.1.0
toml==0.10.2
typing_extensions==4.1.1
